removed loop sum calculations from app layer, using aggr functions in sql

+++ src/main/java/com/shop/demo/service/ProductCategoryService.java
@@ -40,10 +40,10 @@ public class ProductCategoryService {
         return productCategoryDAO.getAllCategory();
     }
 
-    public int profitByCategory(String category1){
+    public long profitByCategory(String category1){
         return productCategoryDAO.profitByCategory(category1);
     }
-    public int revenueByCategory(String category1){
+    public long revenueByCategory(String category1){
         return productCategoryDAO.revenueByCategory(category1);
     }
 }
